requirements.txt
fastapi
uvicorn
tensorflow # Updated TensorFlow version
numpy
pillow
requests
opencv-python
matplotlib
python-multipart